refactor: unify direct tests into 1 file
